% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotated_stores}
\alias{annotated_stores}
\title{annotated list of FHIR stores valid April 2022}
\format{
data.frame
}
\usage{
annotated_stores
}
\description{
annotated list of FHIR stores valid April 2022
}
\keyword{datasets}
