% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{anv_fhir_get_stores}
\alias{anv_fhir_get_stores}
\title{retrieve references to all stores}
\usage{
anv_fhir_get_stores(
  url = test_stores_url(),
  count_str = "1000",
  pyrefs = abfhir_demo()
)
}
\arguments{
\item{url}{character(1) URL for a FHIR store collection}

\item{count_str}{character(1) `_count` value for `where` in ResearchStudy}

\item{pyrefs}{list of Module references (anvil, fhirclient)}
}
\description{
retrieve references to all stores
}
\examples{
allst = anv_fhir_get_stores()
length(allst)
head(sapply(allst, function(x) x$title))
}
