% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{anv_fhir_test_find_patient}
\alias{anv_fhir_test_find_patient}
\title{test extraction of Patient from a given store}
\usage{
anv_fhir_test_find_patient(
  url = anvurl(),
  id = "H_WX-999-999",
  pyrefs = abfhir_demo()
)
}
\arguments{
\item{url}{character(1) URL for a FHIR store}

\item{pyrefs}{list of Module references (anvil, fhirclient)}

\item{count_str}{character(1) `_count` value for `where` in ResearchStudy}
}
\description{
test extraction of Patient from a given store
}
\examples{
t1 = anv_fhir_test_find_patient(url=anvurl("phs002282-DS-CVDRF")) # should be more universal
length(t1)
t1[[1]]$as_json()
}
