% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{anv_fhir_test_list_patients}
\alias{anv_fhir_test_list_patients}
\title{retrieve patients in a ResearchStudy [current limit 1000, needs pagination]}
\usage{
anv_fhir_test_list_patients(
  url = anvurl(),
  count_str = "1000",
  pyrefs = abfhir_demo(),
  as.json = FALSE
)
}
\arguments{
\item{url}{character(1) URL for a store}

\item{count_str}{character(1) `_count` value for `where` in ResearchStudy}

\item{pyrefs}{list of Module references (anvil, fhirclient)}

\item{as.json}{logical(1) convert patient references to JSON before returning}
}
\value{
a list with two elements: first the title of the associated study, then the list of patients
}
\description{
retrieve patients in a ResearchStudy [current limit 1000, needs pagination]
}
\examples{
xx = anv_fhir_test_list_patients(anvurl("phs001272-DS-CSD-MDS"))
names(xx)
xx$title
length(xx$pats)
}
