% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILFHIRSet.R
\name{AnVILFHIRSet}
\alias{AnVILFHIRSet}
\title{constructor}
\usage{
AnVILFHIRSet(urllist, verbose = TRUE)
}
\arguments{
\item{urllist}{list of URLs for FHIR stores}

\item{verbose}{logical(1) messages printed by stages}
}
\description{
constructor
}
