% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.R
\name{consent_abbrev_map}
\alias{consent_abbrev_map}
\title{produce a table of disease/consent code strings in use at dbGAP}
\usage{
consent_abbrev_map(use_dedup = TRUE)
}
\arguments{
\item{use_dedup}{logical(1) defaults to TRUE; otherwise get the table as posted to slack}
}
\description{
produce a table of disease/consent code strings in use at dbGAP
}
\note{
Provided by Chris Wellington of NHGRI in AnVIL slack communication.  The
ad hoc deduplication code is given in inst/tsv/dodedup.R.
}
