% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.R
\name{consent_codes}
\alias{consent_codes}
\title{produce a table of dbGap consent codes, from Table 1 of https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4721915/}
\usage{
consent_codes()
}
\description{
produce a table of dbGap consent codes, from Table 1 of https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4721915/
}
