% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_stores}
\alias{list_stores}
\title{produce vector of names of stores in FHIR server}
\usage{
list_stores()
}
\description{
produce vector of names of stores in FHIR server
}
\note{
uses `anv_fhir_get_stores`
}
\examples{
allst = list_stores()
length(allst)
head(allst)
}
