% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.R
\name{phs2json}
\alias{phs2json}
\title{use FHIR API to get information about a store (possibly resident in AnVIL), using ResearchStudy id}
\usage{
phs2json(phstag = "phs001227")
}
\arguments{
\item{phstag}{character(1) defaults to "phs001227"}
}
\description{
use FHIR API to get information about a store (possibly resident in AnVIL), using ResearchStudy id
}
\examples{
if (interactive()) 
j = phs2json()
j$entry$resource$description
}
