% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.R
\name{phs_desc}
\alias{phs_desc}
\title{use FHIR API to get description from a resource (possibly resident in AnVIL), using ResearchStudy id}
\usage{
phs_desc(x = "phs001227")
}
\arguments{
\item{phstag}{character(1) defaults to "phs001227"}
}
\description{
use FHIR API to get description from a resource (possibly resident in AnVIL), using ResearchStudy id
}
