% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resource_conditions}
\alias{resource_conditions}
\title{acquire information on conditions in a resource}
\usage{
resource_conditions(res, simplify = TRUE)
}
\arguments{
\item{res}{either a resource reference or a list}

\item{simplify}{logical(1) if TRUE, condition data are simplified}
}
\description{
acquire information on conditions in a resource
}
