% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resource_title}
\alias{resource_title}
\title{acquire title for a resource}
\usage{
resource_title(res)
}
\arguments{
\item{res}{either a resource reference or a list}
}
\description{
acquire title for a resource
}
