% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_condition}
\alias{simplify_condition}
\title{very rudimentary helper for rendering condition information}
\usage{
simplify_condition(x)
}
\arguments{
\item{x}{a list assumed to have element `coding` with standard substructure components `display`, `code`}
}
\description{
very rudimentary helper for rendering condition information
}
