% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{store_resources}
\alias{store_resources}
\title{provide information on all resources in a store}
\usage{
store_resources(store = "public", as.json = FALSE, ...)
}
\arguments{
\item{store}{character(1) a store name, defaults to `"public"`}

\item{as.json}{logical(1) apply the `as_json()` method before returning, useful for parallelization}

\item{\dots}{passed to `anv_fire_test_resources`}
}
\description{
provide information on all resources in a store
}
\note{
You can't hand back python objects from worker nodes, so set `as.json = TRUE` in such settings.
}
