% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stores_2022_04}
\alias{stores_2022_04}
\title{list of FHIR stores in AnVIL valid April 2022}
\format{
vector of strings
}
\usage{
stores_2022_04
}
\description{
list of FHIR stores in AnVIL valid April 2022
}
\keyword{datasets}
