% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{anv_fhir_get_stores}
\alias{anv_fhir_get_stores}
\title{retrieve references to all stores}
\usage{
anv_fhir_get_stores(url = test_stores_url(), as.list = TRUE)
}
\arguments{
\item{url}{character(1) URL for a FHIR store collection}

\item{as.list}{logical(1) if TRUE convert JSON to list}
}
\description{
retrieve references to all stores
}
\examples{
allst = anv_fhir_get_stores()
length(allst[[1]])
head(sapply(allst[[1]], function(x) basename(x$name)))
}
