% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMART.R
\name{anvurl}
\alias{anvurl}
\title{long URL string for server announced at https://the-anvil.slack.com/archives/C02HU4GT04B/p1634236625021500}
\usage{
anvurl(tag = "public")
}
\arguments{
\item{tag}{character(1) name of store, defaults to "public"}
}
\description{
long URL string for server announced at https://the-anvil.slack.com/archives/C02HU4GT04B/p1634236625021500
}
